@@ -3,6 +3,9 @@ import { TariffService } from '../../../shared/services/tariff.service';
 import { Tariff } from '../../../shared/models/tariff';
 import { TariffRateComponent } from 'src/app/modules/tariffs/tariff-rate/tariff-rate.component';
 import { TariffRate } from '../../../shared/models/tariff-rate';
+import { Commodity, commodityMap } from 'src/app/shared/models/commodity';
+import { ActivatedRoute } from '@angular/router';
+import { switchMap } from 'rxjs/operators';
 
 @Component({
   selector: 'app-tariffs',
@@ -11,18 +14,22 @@ import { TariffRate } from '../../../shared/models/tariff-rate';
 })
 export class TariffListComponent implements OnInit {
 
-  @ViewChild(TariffRateComponent, {static: false})
+  @ViewChild(TariffRateComponent, { static: false })
   private tariffRateComponent: TariffRateComponent;
 
   tariffs: Tariff[];
   editRowId: number = -1;
 
-  constructor(private tariffsSvc: TariffService) {
+  constructor(private tariffsSvc: TariffService, private route: ActivatedRoute) {
 
   }
 
   ngOnInit() {
-    this.tariffsSvc.getTariffList()
+    this.route.paramMap.pipe(switchMap(p => {
+      let commodity = +p.get('commodity')
+      if (commodity == 0) commodity = 1
+      return this.tariffsSvc.getTariffList(commodity)
+    }))
       .subscribe(data => { this.tariffs = data });
   }
 
@@ -43,15 +50,15 @@ export class TariffListComponent implements OnInit {
     this.tariffRateComponent.openAddDialog(tariff);
   }
 
-  editRate(tariffRate: TariffRate) {
-    const tariff = this.tariffs.find(t => t.id == tariffRate.tariffId);
+  editRate(tariffId:number, tariffRate: TariffRate) {
+    const tariff = this.tariffs.find(t => t.id == tariffId);
     this.tariffRateComponent.openEditDialog(tariff, tariffRate);
   }
-  
+
   deleteRate(tariffId: number, rateId: number) {
     this.tariffsSvc.deleteTariffRate(tariffId, rateId).subscribe(() => {
       const tariff = this.tariffs.find(t => t.id == tariffId);
-      tariff.rates = tariff.rates.filter(r => r.id !== rateId);
+      tariff.rates = tariff.rates.filter(r => r.id != rateId);
     });
   }
 }
