@@ -7,6 +7,5 @@ export const environment = {
   client_secret: '5ee672aa-4daa-4cfc-aafb-30deb8eb42d7',
   apiServer: 'http://10.67.1.238:5050/api/',
   userNotificationHub: 'http://10.67.1.238:5050/user-notification-hub',
-  logoutUri: 'http://10.67.1.238:8080/auth/realms/master/protocol/openid-connect/logout',
-  units: 'м³'
+  logoutUri: 'http://10.67.1.238:8080/auth/realms/master/protocol/openid-connect/logout'
 };
